requirements.txt
addict==2.4.0
apex==0.9.10dev
args==0.1.0
biosppy==2.2.2
data_loading==0.0.1
deepspeed==0.14.2
einops==0.8.0
ftfy==6.2.0
h5py==3.8.0
hydra-core==1.3.2
Jinja2==3.1.2
linear_attention_transformer==0.19.1
lion_pytorch==0.1.4
loguru==0.6.0
matplotlib==3.6.2
mmcv==2.2.0
mmseg==1.3.0
mne==1.3.0
numba==0.59.1
numpy==1.24.1
objgraph==3.6.1
omegaconf==2.3.0
pandas==1.5.3
parse==1.20.1
pavi==0.0.1
Pillow==10.3.0
psutil==5.9.6
pyhealth==1.1.4
python_dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyyaml_include==1.2
regex==2023.10.3
Requests==2.31.0
scikit_learn==1.2.2
scipy==1.13.0
seaborn==0.13.2
setuptools==68.0.0
skimage==0.0
tensorboardX==2.6.2.2
tensorboardX==2.6.2.2
timm==0.3.2
torch==1.13.1
torch_optimizer==0.3.0
torchcde==0.2.5
torchdiffeq==0.2.3
torchsummary==1.5.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.40.2
umap==0.1.1
utilsd==0.0.18.post0
wandb==0.17.0
warmup_scheduler==0.3
